api/requirements.txt
fastapi
pydantic
pandas
scikit-learn
uvicorn
numpy
psycopg2-binary
python-dotenv
sqlalchemy
